fix(scout-for-lol): clear analytics duplication gate
Why: the full CI verification gate rejected the analytics implementation for new copy-paste clones and a stale duplication baseline entry.

What: share Bryan Bucks lifecycle and snapshot aggregation helpers, reuse one analytics test fixture, and remove the obsolete allowed clone from the jscpd baseline.

Verification: bun scripts/jscpd-check.ts; bunx tsc --noEmit --pretty false -p packages/scout-for-lol/packages/backend/tsconfig.json; cd packages/scout-for-lol/packages/backend && bun run test -- src/analytics/bryan-bucks.test.ts src/analytics/product-analytics.test.ts src/analytics/command-usage.test.ts src/analytics/guild-lifecycle.test.ts src/analytics/install-attribution.test.ts
